backend-edited-run-model

+++ backend/cadence/models/RunModel.py
@@ -3,9 +3,9 @@ from .PlaylistModel import Playlist;
 
 class Run(models.Model):
     playlist = models.ForeignKey(Playlist, on_delete=models.CASCADE)
-    avg_pace = models.IntegerField()
+    avg_pace = models.FloatField()
     duration = models.IntegerField()
-    distance = models.IntegerField()
+    distance = models.FloatField()
 
     class Meta:
         db_table = 'runs'
